requirements.txt
boto3~=1.34
dash~=2.14
dash-bootstrap-components~=1.5
dash-core-components~=2.0
dash-html-components~=2.0
dash-mantine-components~=0.12
dash-renderer~=1.9
dash-table~=5.0
python-dotenv~=1.0
